[webdriver]: Add attribution for selenium tests
After speaking with the SFC legal folks, who own the copyright
for the Selenium project, the following change was suggested:

> Under Apache 2.0, you can redistribute modified versions of
> Selenium files (including files ported into a different programming
> language) under a different license, including the W3C's 3-clause BSD
> license.  Be sure to add a note indicating that the files have been
> modified, and include the appropriate copyright/attribution notices (and
> patent/trademark, if applicable) for the original files.

I've added links to the original sources and to the original
Apache 2.0 license, so I think that's now covered properly.
